Changed intialization to use phi instead of threshold
Having reread the paper, using the phi parameter seems to be more
consistent with that was implemented. When query() is called the
threshold defaults to using phi * n_added().

Now use all the keys stored in lhh as the candidate set. Prior we
used just those in the first row. This lead to some keys not making
into the candidate set that should have been, because the collided
in that first row with a key with higher frequency. Using all the
rows reduces this error substantial and doesn't really cost much
time.