v1.4.0 - changed how process_q_item in parallel_add should work
This gives more control to the user to add to the sketches as they
see fit. Should allow for peformance improvements, since you don't
need to return an Iterable or Dict, which parallel_add then would have
to iterate through. This way you can just add directly to the sketches
inside of process_q_item

Updated tests and documentation accordingly.

Also updated hyperloglog.add() so that it can take a value too, though
that value is ignored as it makes no sense to do it for this sketch.